Use `Uint64` for `max_htlc_value_in_flight_msat` consistently
The `max_htlc_value_in_flight_msat` field in the `open_channel` and
`accept_channel` messages is encoded as a `uint64`, while our internal
class for `MilliSatoshi` amounts uses a `Long`. This forced us to store
the remote value as a `UInt64` while our local value was `MilliSatoshi`,
which was inconsistent.

We fix this inconsistency by using an `UInt64` for both the local and
remote values. Since `MilliSatoshi` amounts were already encoded using
64 bits, we can simply switch the codec and we'll be able to start
storing amounts larger than 2^63.

When considering actual HTLC amounts, we can safely round this down to
a `MilliSatoshi` value though, which is useful for channel updates
where the spec disallows setting a value larger than the channel
capacity, where we thus use the `MilliSatoshi` class.

+++ eclair-core/src/test/scala/fr/acinq/eclair/channel/InteractiveTxBuilderSpec.scala
@@ -226,7 +226,7 @@ class InteractiveTxBuilderSpec extends TestKitBaseClass with AnyFunSuiteLike wit
       case (nodeParams, localParams, channelKeys) =>
         RemoteParams(
           nodeParams.nodeId,
-          localParams.dustLimit, UInt64(localParams.maxHtlcValueInFlightMsat.toLong), None, localParams.htlcMinimum, localParams.toSelfDelay, localParams.maxAcceptedHtlcs,
+          localParams.dustLimit, localParams.maxHtlcValueInFlightMsat, None, localParams.htlcMinimum, localParams.toSelfDelay, localParams.maxAcceptedHtlcs,
           channelKeys.revocationBasePoint,
           localParams.walletStaticPaymentBasepoint.getOrElse(channelKeys.paymentBasePoint),
           channelKeys.delayedPaymentBasePoint,
